Implement prefix varint/uvarint coding.
Each value begins with a tag where the number of leading ones indicate
how many following bytes are part of the value. The tag may also contain
part of the value (or all of the value if there are no following bytes).
Values are written in big endian order and occupy 1-9 bytes each.

Encoding and decoding have unsafe implementations that may read up to 9
bytes but are faster if you can meet the invariants. There are also
implementations that extend Buf/BufMut that are safe and use the faster
implementation whenever possible.

The base implementation is unsigned, but signed values may also be
written using a zigzag coding.

Small (single byte) values are fast pathed and inlined and run in the
400-450M elem/s range on my M1 MacBook Air. Longer values are slower to
encode and decode but still achieve ~250M elem/s encoding and 285M
elem/s decoding.