Rework interfaces and improve performance (#6)
Introduce a `PrefixVarInt` trait that applies to all integers sized
16-64 bytes with stand-alone encode and decode methods.
This is not implemented for `usize` or `isize` as those are platform
dependent and callers should use a fixed-size integer
for correctness.

Use the new trait to avoid duplicating methods for unsigned/signed in
interfaces, so this works out roughly like an overload
although turbofish may be required during decoding in some cases. `Buf`
and `BufMut` extensions are retained, but the
`io` support has been turned into stand-alone methods as extending the
interface doesn't make as much sense there -- the
interface exclusively supports ~byte slice.

This change features substantial performance improvements to all
methods, in particular handling of short (single byte)
values is much faster. Computing the encoded length of an integer is
also much faster now.